and even more - refactored but needs debugging
